Add support for schema quotas (#20)
* Add support for schema quotas  (Closes #19)

__Problem__

In mid-2020, Redshift added [quota
support](https://aws.amazon.com/about-aws/whats-new/2020/06/announcing-storage-controls-for-schemas-amazon-redshift/)
to limit the size of schemas. This feature is not yet supported in the
provider. Full details in #19

__Solution__

@coopergillan and I spiked a basic capability. After a few hours of
debugging I verified the code can perform both creates and updates using
quotas. Docs included in solution

The key design idea was that unlimited quotas are specified as strings,
whereas all other quotas are specified as numbers. To get around this,
we can use a fairly common standard of saying 0 = unlimited. There
appears to be no conflicts with other behaviors since 0 is not supported
on Redshift today due to a (at time of authorship) 2048MB lower-limit.
That limit seems to imply that it would be unlikely AWS would ever add a
"0 MB" limit that behaved differently than unlimited.

While GB is the default unit of measure in the documentation, getting
granularity of MB seems to cover all use-cases without having to do
string validation on a string-type parameter: e.g. '25 GB'. That said,
I'm open to the explicit string being a better use-case.

* query restyle

* Add quota limit example and cleanup root.tf