Allows for multiple default grants
Differnt users in redshift can provide multiple overlapping grants to users/groups. This
provides the dumbest quick aggregation to get the right answer, as currently the provider
will only read from the top entry in pg_default_acl, but always write its acl to the entry
tied to the user configured to run the provider, which leads to constant "system changed outside
terraform" warnings that cannot be fixed.